Add support of assignments
